requirements.txt
openai==1.75.0
python-dotenv==1.1.0
requests==2.32.3
google-auth==2.39.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.167.0
google-auth-httplib2==0.2.0
